SceneGraph hierarchy implemented with Rotator component working
About the RotatorComponent: 2 objects; one is moving around a fixed center position and the other object is a child of the first object, moving around its parent (The firts object) at the same time

+++ Minigin/Scene.cpp
@@ -23,7 +23,7 @@ void Scene::Remove(std::shared_ptr<GameObject> object)
 // Remove all "dead" objects after update
 void Scene::RemoveDeadObjects()
 {
-	for (int objectIdx{ 0 }; objectIdx < m_objects.size();)
+	for (size_t objectIdx{ 0 }; objectIdx < m_objects.size();)
 	{
 		if (m_objects[objectIdx]->IsMarkedAsDead())
 		{
